Refactored Code to improve performance and code legibility.
In general the functionality is the same (or similar) to the original PSVG.

The largest change is that Fonts must be defined prior to use in any text object. The font instance is not copied but instead are pointed to from each Text object. This vastly improved performance.

Many of the One-Off classes have not been duplicated at this time (Gradients, Partial Circles, etc) These will be added as necessary.